+++ clients/admin-ui/src/features/common/nav/nav-config.tsx
@@ -177,12 +177,14 @@ export const NAV_CONFIG: NavConfigGroup[] = [
         scopes: [ScopeRegistryEnum.PROPERTY_READ],
       },
       {
-        title: "Messaging",
-        path: routes.MESSAGING_ROUTE,
-        requiresPlus: true,
-        scopes: [ScopeRegistryEnum.MESSAGING_TEMPLATE_UPDATE],
+        title: "Notifications",
+        path: routes.NOTIFICATIONS_ROUTE,
+        scopes: [
+          ScopeRegistryEnum.MESSAGING_TEMPLATE_UPDATE,
+          ScopeRegistryEnum.DIGEST_CONFIG_READ,
+          ScopeRegistryEnum.MESSAGING_CREATE_OR_UPDATE,
+        ],
       },
-
       {
         title: "Privacy requests",
         path: routes.PRIVACY_REQUESTS_SETTINGS_ROUTE,
@@ -290,11 +292,6 @@ export const NAV_CONFIG: NavConfigGroup[] = [
           ScopeRegistryEnum.TCF_PUBLISHER_OVERRIDE_READ,
           ScopeRegistryEnum.TCF_PUBLISHER_OVERRIDE_UPDATE,
         ],
-      },
-      {
-        title: "Messaging providers",
-        path: routes.MESSAGING_PROVIDERS_ROUTE,
-        scopes: [ScopeRegistryEnum.MESSAGING_CREATE_OR_UPDATE],
       },
       {
         title: "About Fides",